Rename SDL-include dir to prevent including files in it accidentally
(instead of the headers provided by the system). This is a part to make FFB working with Linux Mint (22.1 Xia & 22.2 Zara). Next commit will include more related changes.

This will likely break Windows-build, but just use the other branch for it (or fix the include dirs in SConstruct and/or source files).